Modulo_1_Yolo_Vehicle_Detector/requirements.txt
fastapi
uvicorn
opencv-python
pillow
torch
torchvision
ultralytics
python-multipart